requirements.txt
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.8.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.31.0
tiktoken>=0.5.0
